DRTII-1950 Update drt-lib
