+++ build.sbt
@@ -4,7 +4,7 @@ ThisBuild / version := "v" + sys.env.getOrElse("DRONE_BUILD_NUMBER", sys.env.get
 ThisBuild / organization := "uk.gov.homeoffice"
 ThisBuild / organizationName := "drt"
 
-lazy val drtLib = "v1260"
+lazy val drtLib = "v1261"
 
 
 lazy val pekkoVersion = "1.1.3"
